+++ Tests/PlayerTests/Player/MediaSelectionTests.swift
@@ -7,30 +7,11 @@
 @testable import Player
 
 import AVFoundation
-import MediaAccessibility
 import Nimble
 import Streams
 import XCTest
 
-private enum MediaAccessibilityDisplayType {
-    case automatic
-    case forcedOnly
-    case alwaysOn(languageCode: String)
-}
-
 final class MediaSelectionTests: TestCase {
-    private func setupAccessibilityDisplayType(_ type: MediaAccessibilityDisplayType) {
-        switch type {
-        case .automatic:
-            MACaptionAppearanceSetDisplayType(.user, .automatic)
-        case .forcedOnly:
-            MACaptionAppearanceSetDisplayType(.user, .forcedOnly)
-        case let .alwaysOn(languageCode: languageCode):
-            MACaptionAppearanceSetDisplayType(.user, .alwaysOn)
-            MACaptionAppearanceAddSelectedLanguage(.user, languageCode as CFString)
-        }
-    }
-
     func testCharacteristicsAndOptionsWhenEmpty() {
         let player = Player()
         expect(player.mediaSelectionCharacteristics).toAlways(beEmpty(), until: .seconds(2))
@@ -92,7 +73,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testInitialLegibleOptionWithAlwaysOnAccessibilityDisplayType() {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "ja"))
+        MediaAccessibilityDisplayType.alwaysOn(languageCode: "ja").apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.selectedMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("ja"))
@@ -100,7 +81,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testInitialLegibleOptionWithAutomaticAccessibilityDisplayType() {
-        setupAccessibilityDisplayType(.automatic)
+        MediaAccessibilityDisplayType.automatic.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.selectedMediaOption(for: .legible)).toEventually(equal(.automatic))
@@ -108,7 +89,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testInitialLegibleOptionWithForcedOnlyAccessibilityDisplayType() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.selectedMediaOption(for: .legible)).toEventually(equal(.off))
@@ -122,7 +103,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testInitialLegibleOptionWithoutAvailableOptions() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithoutOptions.url))
         expect(player.selectedMediaOption(for: .legible)).toAlways(equal(.off), until: .seconds(2))
@@ -140,7 +121,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testLegibleOptionUpdateWhenAdvancingToNextItem() {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "fr"))
+        MediaAccessibilityDisplayType.alwaysOn(languageCode: "fr").apply()
 
         let player = Player(items: [
             .simple(url: Stream.onDemandWithOptions.url),
@@ -155,7 +136,7 @@ final class MediaSelectionTests: TestCase {
     // this case the perceived selected option must be `.off`.
     @MainActor
     func testLegibleOptionStaysOffEvenIfForcedSubtitlesAreEnabledExternally() async throws {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "ja"))
+        MediaAccessibilityDisplayType.alwaysOn(languageCode: "ja").apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithForcedAndUnforcedLegibleOptions.url))
         await expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
@@ -205,7 +186,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testSelectLegibleOnOption() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
@@ -218,7 +199,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testSelectLegibleAutomaticOption() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
@@ -229,7 +210,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testSelectLegibleOffOption() {
-        setupAccessibilityDisplayType(.automatic)
+        MediaAccessibilityDisplayType.automatic.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
@@ -240,7 +221,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testSelectIncompatibleOptionDoesNothing() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.mediaSelectionOptions(for: .audible)).toEventuallyNot(beEmpty())
@@ -252,7 +233,7 @@ final class MediaSelectionTests: TestCase {
     }
 
     func testLegibleOptionSwitchFromOffToAutomatic() {
-        setupAccessibilityDisplayType(.forcedOnly)
+        MediaAccessibilityDisplayType.forcedOnly.apply()
 
         let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
         expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
@@ -260,76 +241,6 @@ final class MediaSelectionTests: TestCase {
 
         expect(player.selectedMediaOption(for: .legible)).toEventually(equal(.automatic))
         expect(player.currentMediaOption(for: .legible)).to(equal(.off))
-    }
-
-    func testLegibleMediaSelectionCriteriaWithUnknownOrUnavailableLanguage() {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "ja"))
-
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["xy", "it"], for: .legible)
-        expect(player.currentMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("ja"))
-    }
-
-    func testLegibleMediaSelectionCriteriaWithAvailableLanguage() {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "ja"))
-
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["fr"], for: .legible)
-        expect(player.currentMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("fr"))
-    }
-
-    func testLegibleMediaSelectionCriteriaOverride() {
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        expect(player.mediaSelectionOptions(for: .legible)).toEventuallyNot(beEmpty())
-
-        player.select(mediaOption: player.mediaSelectionOptions(for: .legible).first { option in
-            option.languageIdentifier == "ja"
-        }!, for: .legible)
-
-        player.setMediaSelectionCriteria(preferredLanguages: ["fr"], for: .legible)
-        expect(player.currentMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("fr"))
-    }
-
-    func testLegibleMediaSelectionCriteriaReset() {
-        setupAccessibilityDisplayType(.alwaysOn(languageCode: "ja"))
-
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["fr"], for: .legible)
-        expect(player.currentMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("fr"))
-        player.setMediaSelectionCriteria(preferredLanguages: [], for: .legible)
-        expect(player.currentMediaOption(for: .legible)).toEventually(haveLanguageIdentifier("ja"))
-    }
-
-    func testAudibleMediaSelectionCriteriaWithUnknownOrUnavailableLanguage() {
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["xy", "it"], for: .audible)
-        expect(player.currentMediaOption(for: .audible)).toEventually(haveLanguageIdentifier("en"))
-    }
-
-    func testAudibleMediaSelectionCriteriaWithAvailableLanguage() {
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["fr"], for: .audible)
-        expect(player.currentMediaOption(for: .audible)).toEventually(haveLanguageIdentifier("fr"))
-    }
-
-    func testAudibleMediaSelectionCriteriaOverride() {
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        expect(player.mediaSelectionOptions(for: .audible)).toEventuallyNot(beEmpty())
-
-        player.select(mediaOption: player.mediaSelectionOptions(for: .audible).first { option in
-            option.languageIdentifier == "fr"
-        }!, for: .audible)
-
-        player.setMediaSelectionCriteria(preferredLanguages: ["en"], for: .audible)
-        expect(player.currentMediaOption(for: .audible)).toEventually(haveLanguageIdentifier("en"))
-    }
-
-    func testAudibleMediaSelectionCriteriaReset() {
-        let player = Player(item: .simple(url: Stream.onDemandWithOptions.url))
-        player.setMediaSelectionCriteria(preferredLanguages: ["fr"], for: .audible)
-        expect(player.currentMediaOption(for: .audible)).toEventually(haveLanguageIdentifier("fr"))
-        player.setMediaSelectionCriteria(preferredLanguages: [], for: .audible)
-        expect(player.currentMediaOption(for: .audible)).toEventually(haveLanguageIdentifier("en"))
     }
 }
