+++ Tests/PlayerTests/Tools/MediaAccessibilityDisplayType.swift
@@ -1,0 +1,26 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import Foundation
+import MediaAccessibility
+
+enum MediaAccessibilityDisplayType {
+    case automatic
+    case forcedOnly
+    case alwaysOn(languageCode: String)
+
+    func apply() {
+        switch self {
+        case .automatic:
+            MACaptionAppearanceSetDisplayType(.user, .automatic)
+        case .forcedOnly:
+            MACaptionAppearanceSetDisplayType(.user, .forcedOnly)
+        case let .alwaysOn(languageCode: languageCode):
+            MACaptionAppearanceSetDisplayType(.user, .alwaysOn)
+            MACaptionAppearanceAddSelectedLanguage(.user, languageCode as CFString)
+        }
+    }
+}
